+++ regression/do-test.py
@@ -367,7 +367,21 @@ if __name__ == "__main__":
                      "--language", "c++",
                      "--option", "PY_struct_arg=class",
                  ]),
-        TestDesc("vectors"),
+
+        # vectors
+        TestDesc("vectors-numpy", yaml="vectors",
+                 cmdline=[
+                     "--language", "c++",
+                     "--option", "PY_array_arg=numpy",
+                 ]),
+        TestDesc("vectors-list", yaml="vectors",
+                 cmdline=[
+                     "--language", "c++",
+                     "--option", "PY_array_arg=list",
+                     "--option", "wrap_fortran=false",
+                     "--option", "wrap_c=false",
+                 ]),
+        
         TestDesc("forward"),
         TestDesc("example"),
         TestDesc("include"),